Default verdi foedsel
